Fix issues with interpolating (#23)
* split pybind11 wrapped code for easier documenting

* Working on implementing Sphinx/Breathe documentation

* Initial work getting trellis vertex mode sorting up and running

* Closes #21 and #22 plus unrecorded issues

The `InterpolationData<T>` object has been moved to 
`InnerInterpolationData<T>` with a new `InterpolationData<T,R>` object 
holding two independently-typed `InnerInterpolationData<T>` objects. One 
holds the eigenvalues and the other the eigenvectors.
All grid-like objects now have two template parameters for the 
eigenvalue datatype and the eigenvector datatype.

The `PolyhedronTrellis<T,R>` object has been provided with a permutation 
sorting algorithm. This likely still needs refinement.

Previously `InterpolationData<T>` offered a `rotate_in_place()` method 
to transform interpolated values, ID(q_ir), from the irreducible 
reciprocal point, q_ir, to the point related by Q = R^T q_ir + tau. This 
method required all to-be-rotated values to transform like reciprocal 
vectors.
Now `InnerInterpolationData<T>` offers a gateway `rotate_in_place()` 
method to four different transformation functions for values that 
transform like: reciprocal vectors, real space polar vectors, real space 
axial vectors, and (real space) phonon eigenvectors. This has 
necessitated changing the API for ir_moveinto to return pointgroup 
operation *indexes* for R and R^-1, where previously it returned the 3x3 
*values* of R^T and R^-1^T. The brille version number has been 
incremented to highlight this API change.

Supporting the eigevector transformation required adding the possibility 
that `Lattice` objects can have an atom `Basis` and another new class 
has been added to hold the atom permutation and phase-factor vector 
necessary to calculate the transformation.

* Add missing template parameter, fix warnings

Why doesn't MSVC catch missing template parameters in unexecuted code? 
GCC and Clang to the rescue!

+++ src/pointgroup.cpp
@@ -623,7 +623,7 @@ std::vector<std::array<int,9>> get_unique_rotations(const std::vector<std::array
 			if (i_is_not_unique) break;
 		}
 		// if i is unique, add it to the vector of unique rotation indices
-		if (!i_is_not_unique) unique_rot.push_back(i);
+		if (!i_is_not_unique) unique_rot.push_back(static_cast<int>(i));
   }
 
 	// copy from our temporary vector just the unique arrays for output
